Fix: amend Navbar to avoid conflicts
